split RateConverter
